requirements.txt
Django==4.0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
drf-yasg==1.21.5
python-decouple==3.8
psycopg2-binary==2.9.5
gunicorn